LayoutTests/http/tests/site-isolation/navigate-to-about-blank-from-cross-origin-initiated-by-parent-iframe-expected.txt

This test passes if the child frame can modify the grandchild iframe's DOM (once the grandchild iframe has navigated away from a cross-origin site to about:blank).



--------
Frame: '<!--frame1-->'
--------
Child iframe



--------
Frame: '<!--frame2-->'
--------
Hello from your parent
